requirements.txt
langchain
langchain-community
sentence-transformers
faiss-cpu
python-dotenv
google-generativeai
streamlit
streamlit-chat